Move cli_flags.go config.go helpers.go to a new package `config`
